Убрал Autofac из MainForm
