+++ TagsCloud_wpf/App.xaml
@@ -1,8 +1,7 @@
 ﻿<Application x:Class="TagsCloud_wpf.App"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:local="clr-namespace:TagsCloud_wpf"
-             StartupUri="MainWindow.xaml">
+             xmlns:local="clr-namespace:TagsCloud_wpf">
     <Application.Resources>
          
     </Application.Resources>
